mistral-copilot/pyproject.toml
[tool.poetry]
name = "copilot-for-terminal-pro"
version = "0.1.0"
description = ""
authors = ["Michael Struwig <michael.struwig@openbb.finance>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
python-dotenv = "^1.0.1"
httpx = "^0.26.0"
magentic = {extras = ["litellm"], version = "^0.17.0"}


[tool.poetry.group.development.dependencies]
ruff = "^0.1.15"


[tool.poetry.group.dev.dependencies]
ipython = "^8.22.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
